requirements.txt
amqp==5.0.6
appdirs==1.4.4
asgiref==3.4.1
billiard==3.6.4.0
black==21.7b0
celery==5.1.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.5
django-adaptors==0.2.5
djangorestframework==3.12.4
kombu==5.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
prompt-toolkit==3.0.19
pytz==2021.1
redis==3.5.3
regex==2021.7.6
six==1.16.0
sqlparse==0.4.1
tomli==1.1.0
vine==5.0.0
wcwidth==0.2.5
